Enhance Circuit Relay and Discovery Management
- Introduce new structures for relay connection information and relay details, improving the management of relay connections.
- Update the `connect_to_peer` method to check for connected peers using string comparison for peer IDs.
- Implement a new `announce_peers` method in the `DiscoveryManager` to serialize and publish known peers.
- Refactor the `create_relay_transport` function to streamline relay transport creation.
- Update the `NetworkError` enum to include additional error types for better error handling.
- Modify the `PeerInfo` struct to include new fields for agent and protocol versions, enhancing peer data representation.
- Adjust tests to ensure compatibility with the new relay and discovery features.

+++ crates/network/src/metrics.rs
@@ -407,9 +407,9 @@ impl NetworkMetrics {
         self.reputation_changes.with_label_values(&[peer_id, "action"]).inc();
     }
     
-    /// Update a peer's reputation score
+    /// Update the reputation score for a peer
     pub fn update_reputation_score(&self, peer_id: &str, score: i32) {
-        self.reputation_scores.with_label_values(&[peer_id]).set(score);
+        self.reputation_scores.with_label_values(&[peer_id]).set(score as i64);
     }
     
     /// Record that a peer was banned
